test: add CourseRepository Integration Test TaskRepository Integration Test UserRepository Integration Test

+++ src/main/java/br/com/alura/AluraFake/domain/task/rules/general/CourseBuildingStatusTaskRule.java
@@ -21,7 +21,7 @@ public class CourseBuildingStatusTaskRule implements TaskRule {
 
     @Override
     public List<ValidationError> validate(TaskDTO taskRequest) {
-        Optional<Course> course = courseRepository.findById(taskRequest.getCourseId().intValue());
+        Optional<Course> course = courseRepository.findById(taskRequest.getCourseId());
 
         if (!course.isPresent()) {
             return List.of();
